+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/generation/builders/BuiltinFunctionMlogBuilder.java
@@ -21,11 +21,13 @@ public class BuiltinFunctionMlogBuilder extends AbstractFunctionBuilder {
         super(builder);
     }
 
-    public ValueStore handleMlog(AstFunctionCall call, boolean safe, boolean text) {
+    public ValueStore handleMlog(AstFunctionCall call, boolean safe, boolean text, boolean label) {
         if (call.getArguments().isEmpty()) {
             error(call, ERR.FUNCTION_CALL_NOT_ENOUGH_ARGS,
                     call.getFunctionName(), 1, call.getArguments().size());
             return LogicVoid.VOID;
+        } else if (label && call.getArguments().size() != 1) {
+            error(call, ERR.FUNCTION_CALL_WRONG_NUMBER_OF_ARGS, call.getFunctionName(), 1, call.getArguments().size());
         }
 
         assembler.setSubcontextType(AstSubcontextType.ARGUMENTS, 1.0);
@@ -38,6 +40,9 @@ public class BuiltinFunctionMlogBuilder extends AbstractFunctionBuilder {
         final String opcode;
         if (args.getFirst().unwrap() instanceof LogicString str) {
             opcode = str.format(processor);
+            if (label != isLabel(opcode)) {
+                error(args.getFirst(), label ? ERR.MLOG_NOT_A_LABEL : ERR.MLOG_INVALID_OPCODE, call.getFunctionName());
+            }
         } else {
             error(args.getFirst(), ERR.MLOG_FIRST_ARGUMENT_NOT_LITERAL, call.getFunctionName());
             opcode = "noop";
@@ -86,10 +91,14 @@ public class BuiltinFunctionMlogBuilder extends AbstractFunctionBuilder {
             }
         }
 
-        assembler.createCustomInstruction(safe, text, opcode, arguments, parameters);
+        assembler.createCustomInstruction(safe, text, label, opcode, arguments, parameters);
         assembler.setSubcontextType(AstSubcontextType.SYSTEM_CALL, 1.0);
         assembler.clearSubcontextType();
 
         return LogicVoid.VOID;
+    }
+
+    private boolean isLabel(String label) {
+        return label.endsWith(":");
     }
 }