+++ compiler/src/test/java/info/teksol/mc/mindcode/compiler/AbstractTestBase.java
@@ -176,7 +176,11 @@ public abstract class AbstractTestBase {
     }
 
     protected final LogicInstruction customInstruction(String opcode, String... args) {
-        return new CustomInstruction(mockAstContext, false, false, opcode, _logic(args),null);
+        return new CustomInstruction(mockAstContext, false, false, false, opcode, _logic(args),null);
+    }
+
+    protected final LogicInstruction customLabelInstruction(String label) {
+        return new CustomInstruction(mockAstContext, false, false, true, label,List.of(), null);
     }
 
     protected static String q(String str) {